fix utils.ethers, move getLock test to web3Service.ethers
